(style: 캐러셀 UI 완성 (#15)
