--- src/pages/home/view/carousel.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import CarouselTag from "../carouselTag";
-
-function Carousel() {
-  return (
-    <StCarouselBox>
-      <CarouselTag />
-    </StCarouselBox>
-  );
-}
-
-export default Carousel;
-
-const StCarouselBox = styled.div`
-  width: 1000px;
-  margin: 1rem auto;
-  padding: 10px;
-  /* background-color: ${({ theme }) => theme.color.disable}; */
-  display: flex;
-  justify-content: center;
-`;
