+++ src/pages/home/view/index.js
@@ -1,13 +1,12 @@
 import React, { useEffect } from "react";
-import FeedList from "./feedList";
-import Carousel from "./carousel";
+import FeedList from "../feedList";
+import Carousel from "../carousel";
 import { onSnapshot, query } from "firebase/firestore";
 import { postsCollection } from "fb/myfirebase";
 import { useDispatch, useSelector } from "react-redux";
 import { subscribePostsFirestore } from "redux/modules/postsFirestoreState";
 
 export default function Home() {
-  // const [feeds, setFeeds] = useState([]);
   const dispatch = useDispatch();
   // const { posts, loading } = useSelector((modules) => modules.postsFirestoreState);
